+++ src/main/java/com/eclipsekingdom/playerplot/sys/PluginHelp.java
@@ -11,21 +11,21 @@ import org.bukkit.entity.Player;
 
 import java.util.UUID;
 
-import static com.eclipsekingdom.playerplot.sys.Language.*;
-
 public class PluginHelp {
 
     private static String rootCommand = PluginConfig.getRootCommand();
 
     public static void showTo(CommandSender sender) {
-
+        sender.sendMessage("");
         sender.sendMessage(ChatColor.LIGHT_PURPLE + "" + ChatColor.BOLD + "__Player Plot_______");
-        sender.sendMessage(ChatColor.YELLOW + "-------" + ChatColor.GOLD + " " + LABEL_COMMANDS + " " + ChatColor.YELLOW + "-------");
+        sender.sendMessage(ChatColor.GRAY.toString() + "() = " + Language.MISC_OPTIONAL + " " + ChatColor.DARK_PURPLE + "[] = " + Language.MISC_VARIABLE);
+        sender.sendMessage(ChatColor.DARK_PURPLE.toString() + ChatColor.BOLD + "------- " + Language.LABEL_COMMANDS + " -------");
         sendPlotCommands(sender);
-        sender.sendMessage(ChatColor.GOLD + "- - - - - - -");
+        sender.sendMessage(ChatColor.DARK_PURPLE.toString() + ChatColor.BOLD + "------------------");
         sendPlotActionCommands(sender);
-        sender.sendMessage(ChatColor.GOLD + "- - - - - - -");
+        sender.sendMessage(ChatColor.DARK_PURPLE.toString() + ChatColor.BOLD + "------------------");
         sendPlotDeedCommand(sender);
+        sender.sendMessage("");
     }
 
     public static void showPlots(Player player) {
@@ -34,11 +34,14 @@ public class PluginHelp {
         PermInfo permInfo = UserCache.getPerms(playerID);
         int used = PlotCache.getPlayerPlotsUsed(playerID);
         int capacity = PluginConfig.getStartingPlotNum() + permInfo.getPlotBonus() + userData.getUnlockedPlots();
-        player.sendMessage(ChatColor.LIGHT_PURPLE + "" + ChatColor.BOLD + LABEL_PLOTS + ChatColor.ITALIC + "" + ChatColor.DARK_PURPLE + " -  (" + used + "/" + capacity + ")");
-        player.sendMessage(ChatColor.YELLOW + "-------" + ChatColor.GOLD + " " + LABEL_COMMANDS + " " + ChatColor.YELLOW + "-------");
+        player.sendMessage("");
+        player.sendMessage(ChatColor.LIGHT_PURPLE + "" + ChatColor.BOLD + Language.LABEL_PLOTS + ChatColor.ITALIC + "" + ChatColor.DARK_PURPLE + " -  (" + used + "/" + capacity + ")");
+        player.sendMessage(ChatColor.GRAY.toString() + "() = " + Language.MISC_OPTIONAL + " " + ChatColor.DARK_PURPLE + "[] = " + Language.MISC_VARIABLE);
+        player.sendMessage(ChatColor.DARK_PURPLE.toString() + ChatColor.BOLD + "------- " + Language.LABEL_COMMANDS + " -------");
         sendPlotCommands(player);
-        player.sendMessage(ChatColor.GOLD + "- - - - - - -");
+        player.sendMessage(ChatColor.DARK_PURPLE.toString() + ChatColor.BOLD + "------------------");
         sendPlotActionCommands(player);
+        player.sendMessage("");
     }
 
 
@@ -47,29 +50,31 @@ public class PluginHelp {
     }
 
     private static void sendPlotCommands(CommandSender sender) {
-        sendCommand(sender, "&6/" + rootCommand + " scan: &f" + HELP_PLOT_SCAN.toString());
-        sendCommand(sender, "&6/" + rootCommand + " claim &7([" + ARG_PLOT + "])&6: &f" + HELP_PLOT_CLAIM.toString());
-        sendCommand(sender, "&6/" + rootCommand + " list: &f" + HELP_PLOT_LIST.toString());
-        sendCommand(sender, "&6/" + rootCommand + " flist: &f" + HELP_PLOT_FLIST.toString());
-        sendCommand(sender, "&6/toplot &c[plot]&6: &f" + HELP_TOPLOT.toString());
+        sendCommand(sender, "&d/" + rootCommand);
+        sendCommand(sender, "&d  > scan: &f&o" + Language.HELP_PLOT_SCAN.toString());
+        sendCommand(sender, "&d  > claim &7([" + Language.ARG_PLOT + "])&d: &f&o" + Language.HELP_PLOT_CLAIM.toString());
+        sendCommand(sender, "&d  > list: &f&o" + Language.HELP_PLOT_LIST.toString());
+        sendCommand(sender, "&d  > flist: &f&o" + Language.HELP_PLOT_FLIST.toString());
+        sendCommand(sender, "&d/toplot &5[plot]&d: &f&o" + Language.HELP_TOPLOT.toString());
     }
 
 
     private static void sendPlotActionCommands(CommandSender sender) {
-        sendCommand(sender, "&6/" + rootCommand + " &7(@[" + ARG_PLOT + "])&6 rename &c[" + ARG_NAME + "]&6: &f" + HELP_PLOT_RENAME.toString());
-        sendCommand(sender, "&6/" + rootCommand + " &7(@[" + ARG_PLOT + "])&6 free: &f" + HELP_PLOT_FREE.toString());
-        sendCommand(sender, "&6/" + rootCommand + " &7(@[" + ARG_PLOT + "])&6 info: &f" + HELP_PLOT_INFO.toString());
-        sendCommand(sender, "&6/" + rootCommand + " &7(@[" + ARG_PLOT + "])&6 trust &c[" + ARG_PLAYER + "]&6: &f" + HELP_PLOT_TRUST.toString());
-        sendCommand(sender, "&6/" + rootCommand + " &7(@[" + ARG_PLOT + "])&6 untrust &c[" + ARG_PLAYER + "]&6: &f" + HELP_PLOT_UNTRUST.toString());
-        sendCommand(sender, "&6/" + rootCommand + " &7(@[" + ARG_PLOT + "])&6 upgrade: &f" + HELP_PLOT_UPGRADE.toString());
-        sendCommand(sender, "&6/" + rootCommand + " &7(@[" + ARG_PLOT + "])&6 downgrade: &f" + HELP_PLOT_DOWNGRADE.toString());
-        sendCommand(sender, "&6/" + rootCommand + " &7(@[" + ARG_PLOT + "])&6 setcenter: &f" + HELP_PLOT_SET_CENTER.toString());
-        sendCommand(sender, "&6/" + rootCommand + " &7(@[" + ARG_PLOT + "])&6 setspawn: &f" + HELP_PLOT_SET_SPAWN.toString());
+        sendCommand(sender, "&d/" + rootCommand + " &7(@[" + Language.ARG_PLOT + "])");
+        sendCommand(sender, "&d  > rename &5[" + Language.ARG_NAME + "]&d: &f&o" + Language.HELP_PLOT_RENAME.toString());
+        sendCommand(sender, "&d  > free: &f&o" + Language.HELP_PLOT_FREE.toString());
+        sendCommand(sender, "&d  > info: &f&o" + Language.HELP_PLOT_INFO.toString());
+        sendCommand(sender, "&d  > trust &5[" + Language.ARG_PLAYER + "]&d: &f&o" + Language.HELP_PLOT_TRUST.toString());
+        sendCommand(sender, "&d  > untrust &5[" + Language.ARG_PLAYER + "]&d: &f&o" + Language.HELP_PLOT_UNTRUST.toString());
+        sendCommand(sender, "&d  > upgrade: &f&o" + Language.HELP_PLOT_UPGRADE.toString());
+        sendCommand(sender, "&d  > downgrade: &f&o" + Language.HELP_PLOT_DOWNGRADE.toString());
+        sendCommand(sender, "&d  > setcenter: &f&o" + Language.HELP_PLOT_SET_CENTER.toString());
+        sendCommand(sender, "&d  > setspawn: &f&o" + Language.HELP_PLOT_SET_SPAWN.toString());
     }
 
 
     private static void sendPlotDeedCommand(CommandSender sender) {
-        sendCommand(sender, "&6/plotdeed &c[" + ARG_PLAYER + "] [" + ARG_AMOUNT + "]&6: &f" + HELP_PLOT_DEED.toString());
+        sendCommand(sender, "&d/plotdeed &5[" + Language.ARG_PLAYER + "] [" + Language.ARG_AMOUNT + "]&d: &f&o" + Language.HELP_PLOT_DEED.toString());
     }
 
 }