+++ CHANGELOG.md
@@ -5,6 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.19.0] - ???
+
+### Added
+
+- Added introductory message to /playerplot command
+- Added /playerplot info command
+
+### Changed
+
+- Improved help command formatting
+- Changed help command styling to match plugin theme
+
 ## [1.18.0] - 2020-09-14
 
 ### Added